Support generating representation structs
